Fix processing of DW_AT_ranges for DWARF 5
In DWARF 5, the "form" of the attribute of DW_AT_ranges can still be
DW_FORM_sec_offset, or the new DW_FORM_rnglistx. In DWARF 5,
DW_AT_ranges is always a rnglist (ie, points into dwarf_rnglists),
rather than a rangelistptr (points into dwarf_ranges).

The DW_FORM_sec_offset is now an offset into dwarf_rnglists, while
DW_FORM_rnglistx is an index into the offset table in the rnglists
section - this indirection makes it easier to deal with relocations, but
the DW_FORM_sec_offset is still used by gcc.

The previous iteration of the code decided, based on the form, to use
either the old "rangelistptr" class, or the new "rnglist" class - that's
wrong - instead, always use the new rnglists code, but skip the
indirection part for DW_FORM_sec_offset.

The relevant text from the DWARF 5 spec are on pages 210 and 216:

Page 210 shows the valid classes for each attribute, and includes:
```
DW_AT_ranges 0x55 rnglist
```

Page 216 describes `rnglist`:
```
0  rnglist
11 This is represented as either:
12 - An index into the .debug_rnglists section (DW_FORM_rnglistx). The
13 unsigned ULEB operand identifies an offset location relative to the
14 base of that section (the location of the first offset in the section, not the
15 first byte of the section). The contents of that location is then added to
16 the base to determine the location of the target range list of entries.
17 - An offset into the .debug_rnglists section (DW_FORM_sec_offset).
18 The operand consists of a byte offset from the beginning of the
19 .debug_rnglists section. It is relocatable in a relocatable object file,
20 and relocated in an executable or shared object file. In the 32-bit
21 DWARF format, this offset is a 4-byte unsigned value; in the 64-bit
22 DWARF format, it is an 8-byte unsigned value (see Section 7.4 on
23 page 196).
24 This class is new in DWARF Version 5
```

Now that this actually works, we find the various
previously-unimplemented directives in the rnglists - so implement
those.